[win10] peripherals: move implementation to platform folder.
