+++ xbmc/platform/win10/peripherals/CMakeLists.txt
@@ -2,4 +2,4 @@ set(SOURCES PeripheralBusUSB.cpp)
 
 set(HEADERS PeripheralBusUSB.h)
 
-core_add_library(peripherals_bus_win10)
+core_add_library(platform_win10_peripherals)
