+++ xbmc/platform/win10/peripherals/PeripheralBusUSB.cpp
